+++ src/unixfs/get.rs
@@ -1,4 +1,4 @@
-use std::{path::Path, time::Duration};
+use std::{path::Path, time::Duration, num::NonZeroU8};
 
 use either::Either;
 use futures::{future::BoxFuture, stream::BoxStream, FutureExt, Stream, StreamExt};
@@ -18,6 +18,7 @@ pub fn get<'a, P: AsRef<Path>>(
     providers: &'a [PeerId],
     local_only: bool,
     timeout: Option<Duration>,
+    retry: Option<NonZeroU8>,
 ) -> UnixfsGet<'a> {
     let (repo, dag, session) = match which {
         Either::Left(ipfs) => (
@@ -52,7 +53,7 @@ pub fn get<'a, P: AsRef<Path>>(
             };
 
         let block  = match dag
-            .resolve_with_session(session, path.clone(), true, providers, local_only, timeout, None)
+            .resolve_with_session(session, path.clone(), true, providers, local_only, timeout, retry)
             .await
             .map_err(TraversalFailed::Resolving)
             .and_then(|(resolved, _)| resolved.into_unixfs_block().map_err(TraversalFailed::Path)) {
@@ -70,7 +71,7 @@ pub fn get<'a, P: AsRef<Path>>(
 
         while walker.should_continue() {
             let (next, _) = walker.pending_links();
-            let block = match repo.get_block_with_session(session, next, providers, local_only, timeout, None).await {
+            let block = match repo.get_block_with_session(session, next, providers, local_only, timeout, retry).await {
                 Ok(block) => block,
                 Err(e) => {
                     yield UnixfsStatus::FailedStatus { written, total_size, error: Some(anyhow::anyhow!("{e}")) };