+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 0.10.0
+- feat: Retry fetching block after timeout. [PR XXX](https://github.com/dariusc93/rust-ipfs/pull/XXX)
+
 # 0.9.0
 - feat: Add libp2p-bitswap and feature gate both implementation. [PR 126](https://github.com/dariusc93/rust-ipfs/pull/126)
 - feat: Add `Repo::pin`, `Repo::remove_pin` and expanded options [PR 122](https://github.com/dariusc93/rust-ipfs/pull/122)